Semi-clearing screen support added.
It is still buggy, and has some issues that need to be further investigated with not showing the correct page. Also, the code is far from efficient or elegant. This is more of a hacked-together proof-of-concept solution to the problem. The solution has been found, and it is in the works, this is a beta version of what has been created so far. Proceed with caution on this commit. It may have also inadvertently broken other parts of the code as well.