Switch the order of the audio formats on firefox playback
Firefox has a bug where it can't seek in the audio-only webm files with
no cues, and it seems like they have no intention of fixing this...
Serve up the ogg files first for them.

Adding cues to audio-only webm files is *hard*, there are no standard
tools that support doing this cleanly.