Add a script to handle updating links in existing recordings
This adds the version number to the playback links so that
recordings select the correct version-specific playback support
files.

This script may be run automatically during upgrade; in that
case it should be run like

.../bbb-0.9-beta-recording-update --quiet

After copying recordings from an old BigBlueButton server, you
may want to manually re-run the script, and it takes the option
--force to recheck all recordings even if it has previously been
run.