Fix pan and zoom in the responsive playback page
When pan/zoom was done, the image would zoom and move inside the svg, that
occupied a larger area than the viewbox we wanted to show.

Had to control the size of the container that has the svg via javascript,
to make it exactly the size we want to display the presentations and with
this crop all parts of the presentation that lie outside the container.