Fix playback of recordings not loading on firefox mobile
The event "loadeddata" is only triggered on firefox mobile once the
video/audio is played, so it requires the user to click on the play button
for it to be triggered.
The playback page was waiting for the event to be triggered before
removing the "loading" page, and that would never happen on firefox mobile.

Now we don't wait anymore for the video to be loaded, we just wait for
the slides and the initial setup.